folder restructuring

+++ src/modules/data_pipeline.py
@@ -1,14 +1,18 @@
-from zmq import PROTOCOL_ERROR_ZMTP_MALFORMED_COMMAND_UNSPECIFIED
-from helper_functions import segment_cough, preprocess_cough, classify_cough, label_func
-from coughvid_functions import classify_cough
-from xgboost import XGBClassifier
+import os
+import librosa
+import numpy as np
+from PIL import Image
+from joblib import load
 from scipy.io import wavfile
 from fastai.vision.all import *
-from joblib import load
-from PIL import Image
-import numpy as np
-import librosa
-import os
+from xgboost import XGBClassifier
+from modules.coughvid_functions import classify_cough
+from modules.helper_functions import (
+    segment_cough,
+    preprocess_cough,
+    classify_cough,
+    label_func
+)
 
 import warnings
 warnings.filterwarnings('ignore') 
@@ -24,7 +28,7 @@ def check(model_dir, audio_path):
     return probability
 
 
-def process(path, save=True, display=False):
+def process(path, save=True, mkdir=False, display=False):
     images = []
     audio, rate = librosa.load(path)
     segments, _ = segment_cough(audio, rate)
@@ -38,12 +42,14 @@ def process(path, save=True, display=False):
             if display:
                 img.show(img)
     if save:
-        cache_path = os.path.join(path[:path.rindex('/')], 'cache')
-        if not os.path.exists(cache_path):
-            os.mkdir(cache_path)
-        else:
-            for file in os.listdir(cache_path):
-                os.remove(os.path.join(cache_path, file))
+        cache_path = path[:path.rindex('/')]
+        if mkdir:
+            cache_path = os.path.join(path[:path.rindex('/')], 'cache')
+            if not os.path.exists(cache_path):
+                os.mkdir(cache_path)
+            else:
+                for file in os.listdir(cache_path):
+                    os.remove(os.path.join(cache_path, file))
         for i, img in enumerate(images):
             img.save(os.path.join(cache_path, f'{i}.png'))
     return images
@@ -73,7 +79,8 @@ def pipeline(
     cnn_model_path,
     check_threshold=0.8,
     predict_threshold=0.8,
-    weighing_formula='average'
+    weighing_formula='average',
+    verbose=False
     ):
     audio_path = os.path.join(
         data_directory,
@@ -85,13 +92,13 @@ def pipeline(
     # check if the sample contains cough
     check_probability = check(xgb_model_path, audio_path)
     if check_probability < check_threshold:
-        return 0 # no cough
+        return 'NA' # no cough
     # process the sample
     process(audio_path)
     # predict the sample
     cache_path = os.path.join(data_directory, 'cache')
     cnn_model_path = os.path.join(cnn_model_path, 'model.pkl')
-    predictions = predict(cnn_model_path, cache_path)
+    predictions = predict(cnn_model_path, cache_path, verbose=verbose)
     if weighing_formula == 'average_floor':
         prediction = int(np.floor(np.average(predictions)))
     elif weighing_formula == 'average_ceil':
@@ -103,11 +110,11 @@ def pipeline(
         predictions = [1 if el > predict_threshold else 0 for el in predictions]
         prediction = int(np.ceil(np.average(predictions)))
     return prediction
-    print(prediction)
+    # print(prediction)
 
 
 
-pipeline('../trials', '../models/xgb', '../models/cnn', weighing_formula='average_ceil')
-pipeline('../trials', '../models/xgb', '../models/cnn', weighing_formula='average_floor')
-pipeline('../trials', '../models/xgb', '../models/cnn', weighing_formula='majority_ceil')
-pipeline('../trials', '../models/xgb', '../models/cnn', weighing_formula='majority_floor')
+# pipeline('../trials', '../models/xgb', '../models/cnn', weighing_formula='average_ceil')
+# pipeline('../trials', '../models/xgb', '../models/cnn', weighing_formula='average_floor')
+# pipeline('../trials', '../models/xgb', '../models/cnn', weighing_formula='majority_ceil')
+# pipeline('../trials', '../models/xgb', '../models/cnn', weighing_formula='majority_floor')
